Combine sidebars for TFC/TFE
TFC and TFE are the same application, but we currently don't have the tooling to
serve two totally separate versions of the application docs without an
unbearable maintenance burden. Our solution thus far has been to use the TFC
docs as the app docs, and the TFE docs as the deployment/operation docs for
running an instance.

OK in theory, but in practice it makes the docs frustrating for the TFE
application users. This change is intended to:

- Help with discoverability by making the entire TFC docs nav hierarchy visible
within the TFE deployment/ops docs.
- Help with confirmation/belief by changing the TFC sidebar header to
"Terraform Cloud / Terraform Enterprise", so TFE users can trust that
they're in the right place.